split file and use opt package
