+++ kadai2/matsuyoshi30/opt/opt_test.go
@@ -1,4 +1,4 @@
-package main
+package opt
 
 import (
 	"testing"
